distgit: extend the timeout limit for importing packages
Imporing should be fast for the majority of packages. It is basically
instant for packages with one or couple of sources. I am extending the
timeout limit just because of texlive package which has ~6k
sources.

There is probably a space for improvement in the importing code so we
shouldn't extend the timeout infinitely but rather rework the
code. For the time being, increasing the timeout to three hours should
be good enough.