frontend: don't re-set Build.package value
When the package is already assigned to build, we don't want to
re-assign it to a different package - even if the source RPM had a
different name (pkg_name in /update/ route payload).  The `Package`
object basically only holds a concrete build configuration, and so
picking a different configuration means that the build may suddenly
start behaving against the user's expectation (e.g. the chroot selected
chroot list we build into may be different).

Spotted in PR#1530 by @schlupov.