frontend: hide buttons in repositories pages for non-admins
Since we take the permissions into account when handling the POST data,
it is safe to just hide the button to not cause confusion.  The other
option is to not even list the projects where user has builder
permissions, but it looks to be wise to show it - so the builder user
can contact the real administrators.