beaker-tests: fix timeout tests
There are couple of fixes for the timeout tests. First, we don't need
to `copr watch-build` separately since we do not run the build with
`--nowait`. Also, when the build is expected to fail because of a low
timeout setting, we need to expect a non-zero status code.

Lastly some magic regarding curl-ing the log, I just refactored it a
bit so it is easier to debug, it doesn't fail now, but I don't really
know what whas wrong with the previous version.

Merges: #1552