frontend: allow searching for build ID
Fix #516

This feature is the same in e.g. Bugzilla. When you search for a
whatever string, you will get a list of the results. When you search
for a number it is considered to be a build ID and you will get
redirected to its detail.

If a build with searched number doesn't exist, we proceed with
fulltext searching the number as it was before introducing this feature.