frontend: not allow to extend already expired chroots
It's the safer option. Otherwise we could have already created
delete action for a chroot (just not yet finished because of a big
queue or something) and user could change his mind and extend. Then
the action would be processed and the chroot removed anyway. Let's
just not allow user to extend such chroot in the first place. We give
users plenty of time (half a year) so we don't have to deal with this
corner case regarding a few hours window after a chroot is aleady
expired but not yet deleted.

However, by doing this we would disallow users to change their mind if
they intentionally click to expire a chroot and then want to take it
back. This is a valid use-case and we want to keep it, so we decided
to change the semantics of "Expire" button from "by this moment, the
chroot is expired" to "this chroot is going to expire in X hours". By
default, this is going to be 12 hours. That is long enough to
potentially extend the chroot again and at the same time good enough
to not bother user with any more notifications and to reclaim the disk
space shortly.