frontend: use URL path parameter (if there is such) instead of failing
After 5bfd679, the following route stopped working

    @apiv3_ns.route("/build-chroot/build-config/<int:build_id>/<chrootname>",
                    methods=GET)

and started throwing The this error

    ipdb> response.json
    {'error': 'Missing argument build_id'}

This was caused by decorating the `get_build_chroot_config` view with
`@query_params` decorator. Let me elaborate. The purpose of
`@query_params` is to allow defining views like this:

    @apiv3_ns.route("/build-chroot/build-config", methods=GET)
    @query_params()
    def get_build_chroot_config(build_id, chrootname):

Then it will ensure the `/build-chroot/build-config` URL is called
with `build_id=foo` and `chrootname=bar` query parameters, i.e.

    /build-chroot/build-config?build_id=foo&chrootname=bar

and then pass the values to the respective function arguments.

The `@query_parameters` decorator wasn't invented with the possibility
of URL path parameters and query parameters with same names in mind. I
couldn't imagine how it could be useful. Well, the 5bfd679 found a
first use-case, providing a route with query parameters while
keeping an old route with the same path parameters (for preserving
backward compatibility).

    @apiv3_ns.route("/build-chroot/build-config", methods=GET)
    @apiv3_ns.route("/build-chroot/build-config/<int:build_id>/<chrootname>",
        methods=GET)  # deprecated
    @query_params()
    def get_build_chroot_config(build_id, chrootname):

This commit changes `@query_params` to not fail for the old route
because of missing query parameters and rather use the parameter from
URL path (i.e. do nothing for that parameter, the value of path
parameter is assigned to the view function argument by default in Flask).

I have also added one more test to check that every variant works.