doc: display copr logo in the panel
I needed to add a new logo with transparent background because our
Fedora blue color doesn't match the RTD theme blue.

To resize the logo in panel, change the canvas ratio of the
image. Basically the wider the (transparent) background is, while
keeping the same height, the smaller it shows in the panel.