rpmbuild: don't override isolation config by default
The default --isolation in Mock is 'auto'.  But e.g. epel-6 has
the isolation explicitly set to 'simple' (because 'nspawn' is known to
not work correctly).

So doing isolation=auto in copr-rpmbuild for the "default" value would
effectively mean that we are switching to isolation=auto also the epel-6
in Copr (which would in turn mean we would use non-working
isolation=nspawn).

Fixes: #1603