cli: fix serializable() helper with new python3-munch
The Munch(List(items=[])) structure isn't possible to copy().  The
documentation for Munch.copy() says it does shallow copy (which doesn't
seem to be truth anymore, otherwise they shouldn't call the constructor
for sub-arrays).

The problem is of course on our side, because it is terribly ugly to
inherit from list() and change it's default behavior (the required items
argument).

Since we indeed want a shallow copy, doing just dict() should be enough.
We loose the Munch() wrapper, but the output from serializable() isn't
used for any purpose where Munch would be needed.

Fixes: rhbz#1899274