backend: pruner: simplify the pool logic
Callbacks in apply_async() are unnecessary, we don't have to take care
if run_createrepo() finishes correctly or not - we just log the failure
and worker goes on..

The arguments passed down to run_createrepo() were duplicated in "cmd".
So construct the called command down in run_createrepo() instead.

Drop the __del__() method which is not guaranteed to be called.

Don't ignore the signal;  because we don't implement our own signal
handler, we definitely want to deliver CTRL+C to all processes in this
process group (including createrepo_c or prunerepo).