frontend: fix chroots in resubmit form after b3af506
For some reason this issue didn't happen in local dev environemnt
because of another bug when `copr.active_chroots` is a `filter` so
calling `list(copr.active_chroots)` for the second time returned empty
list, hence later in this function `available_chroots` was empty. I am
fixing it as well.