frontend: rework chroots fields in copr forms (take 2)
See PR#1656 and temporary revert in PR#1662

This is a follow-up to b3af506c5 doing the exact same rework, but now
for copr form. Please see b3af506c5 commit message for more information.

This is a second take of that change, see PR#1662 for why it was
reverted the first time.

This time I run beaker tests against this change and made sure they
pass. To check that APIv1 still works, I updated the
`test_api_create_copr_ok_all` test to create a project with multiple
chroots (which previously was an issue). Unfortunatelly we don't have
any test for APIv1 project edit, so I tested just

    chroots = ["fedora-33-x86_64", "fedora-rawhide-x86_64"]
    client = CoprClient.create_from_file_config()
    client.modify_project("foo", chroots=chroots)