backend: pruner: limit number of concurrent jobs
Default to `os.cpu_count(), from the multiprocessing.Pool defaults.  The
previous value 50 made the load too high both on production and devel
instance.

While on it, I'm adding the `prune_workers` config variable to config
parser.

I renamed the variable from `max_prune_processes` to `prune_workers` as
that matches the limit better (first the final number of spawned
processes is much higher as the workers spawn other processes, and
second it is rather static number, not maximal number of workers).