doc: add IPv6 instructions for upgrading persistent VMs
We need to assign IPv6 address right before we spawn the infrastructure
machine.  If this is done immediately, the IPv6 networking is properly
configured automatically on the host by cloud-init.  But if we assign
the IPv6 later - manual IPv6 configuration and reboot is needed.

Because we always assign new IPv6 to new infra machines, I'm adding also
a new step for fixing IPv6 DNS records.