frontend: add base form for creating and modifying projects
This is a follow-up after discussion in PR#1676. We don't want to
copy-paste the same attributes to copr creation form and
`CoprModifyForm`, so I am suggesting to have a base class for them.

Actually, it seems like `CoprModifyForm` is unnecessary and we can use
the newly created `CoprForm` directly.

Some fields such as `persistent`, `delete_after_days`, and `chroots`
needs to be overwritten within the `CoprFormFactory` because they
depend on a `copr` instance that is passed to the factory.