frontend: better identify the build submitter
Major motivation
https://pagure.io/pagure/issue/4569

The problem is that the 'agent' field isn't set to the user who
actually triggered the fedora-messaging event, but to artificial
string "pagure" (a non-existing user e.g. in pagure.io).

Therefore we have to look at better data, which is in the
pullrequest.user.name field.  This field is not in the "push" event
so we keep using the "agent" there (in this case it is set correctly).