frontend: introduce ChrootDeletionStatus
When a chroot is marked as EOL or when it is unclicked from a project,
it goes through several stages before its data is finally deleted.
We don't store this state to database but calculate it from several
other properties. Defining a proper enum should help us not having
hack all over the place but rather condense it within one well-tested
method.