frontend: catch tracebacks when rendering invalid modules
Fix #1752

We have some modules with invalid modulemd YAML files in our
database. The validation used to be rather benevolent in the initial
implementations. With this, we can print a user-friendly for such
modules. At this moment, there are 57 modules in the Fedora Copr
instance, that cannot be parsed. One of them because of

    Hint: Unexpected scalar outside of map. [line 71 col 11]

The rest of them because of

    Hint: Module license is missing