frontend: hide deactivated chroots in the project overview
Fix RHBZ 1931377

AFAIK we don't want to list deactivated (not unclicked but rather
deactivated by Copr administrators) chroots in a project overview
because it confuses the project owners (they don't see such chroots in
project settings and therefore have no way to hide them in the
overview).

On the other hand, AFAIK we want to list EOL chroots, that are not
deleted yet, in the project overview. They will automatically
disappear once their data is deleted (which can be done promptly in
the /repositories page).

What chroots we want to display is covered by the
`test_permissible_chroots` test, which should describe all the
possible cases.