frontend: use correct auto_prune default when creating via API
Fix #1747
Fix #1746

The `test_update_copr_api3` test used web UI to create a new
project. I changed it to API, which revealed that
`follow_fedora_branching` and `runtime_dependencies` values weren't
set at all, so I fixed it and updated the test.

The behavior of `default` form values is terribly confusing to me, I
described my observations in the `set_defaults` function. It might
be wrong though.

Also, please note that the #1747 issue cannot be reproduced via

    copr-cli create test-1 --chroot fedora-33-x86_64

and neither it can't be reproduced via

    from copr.v3 import Client
    Client.create_from_config_file()
    client.project_proxy.add("frostyx", "test-2", ["fedora-33-x86_64"])

because both of them send a complete `CoprForm` data. It can be
reproduced only the way this test works, i.e. sending a subset of
`CoprForm` to `/api_3/project/add/<NAME>`.