builder, backend, frontend, python: provide build results via APIv3
Fix #1411

This change is not trivial. We need to generate `results.json` file
containing the the built packages NEVRAs and we then download this
file to backend. When build is marked as finished, we send these
results to the frontend and store them to the database.

The APIv3 endpoints are just a tip of the iceberg. We also provide
`get_built_packages(...)` methods in our python client but we don't
provide any `copr-cli` command (yet?) that would utilize them.