cli: migrate from CoprClient APIv1 call for generating webhook secrets
See #1792

As it turns out, we still use APIv1 for the following command

    copr new-webhook-secret <owner>/<project>

Since we don't have this feature implemented in APIv3 yet, I would
prefer not doing so in the pre-release window, and rather exploit the
fact that we didn't drop APIv1 support from frontend and send the
request directly.

APIv3 endpoint should be implemented and this command should be
migrated ASAP after the release.