backend: log the partial "prunerepo" actions
It may happen that we remove only a part of a subdirectory by repo
pruning cron job.  Previously the files silently disappeared, now we at
least log when it happened, and what PID did that (so Copr
administrators can search through the pruner.log and modifyrepo.log
to find more info).

Merges: #1806
Relates: #1804