frontend: fix unclicking chroots from project settings
We have an issue, that we _can't_ disable a chroot from project
settings. Reproducer:

1. Go to a project settings
2. Unclick some chroot
3. Click on "Update" to submit the form
4. Go to the project settings again
5. The chroot is not unclicked

In reality, it properly disables the chroot in the database, it is
only displayed wrong in the form. This commit fixes it.

It is still displayed in the project overview though but that is AFAIK
a feature, not a bug. But maybe we should show some additional
information there, that the chroot repo is still available and
therefore displayed but project admin requested its deletion?