frontend: show deletion tooltip in project overview
In the project overview, we currently show all available repositories
as opposite to just chroots that are enabled in the project
settings. On top enabled chroots, also EOL chroots and chroots that
were disabled by the project owner, that are not expired yet.

This is a kind step towards people that use systems that are not supported
anymore but their repositories haven't been shutdown yet. However, it
can be a bit confusing for project owners because they might e.g. go
to the project settings, unclick some chroot, submit the form, which
will redirect them to the project overview and the same chroot is
there for some reason. And yet they cannot see it enabled in the
project settings, ...

I am adding a small icon indication that the chroot is there
temporarily and that it is going to be deleted in XY days.

Unfortunatelly such a simple change required a bit of refactoring
because when generating the repositories, we operated over
mock_chroots instead of copr_chroots and therefore there wasn't any
project-specific information available.