frontend, backend: allow pruning also the finalized chroots
Since we had a lot of problems with copr_prune_results.py recently, we
should at least once re-run the prunerepo script against all the chroots
we keep in resultdir - even though we already did set the
final_prunerepo_done flag on the Frontend side.

For this purpose, a new option --prune-finalized-chroots (that implies
--no-mtime-optimization) comes in handy.

To correctly implement this, though, we need to give the Backend side
the **full list** of MockChroots we have in database (also deactivated
ones) and the info both info (a) whether the final prunerepo was done,
and (b) if the chroot is actually active.  The list of chroots is going
to be larger and larger over the years so we might want to optimize this
in the future (though at the time of pushing this task we run the query
only once a day).