cli: api: frontend: GSSAPI support
The GSSAPI configuration is opt-in in python-copr code, but enabled by
default in copr-cli code.  Authentication using a token file
(.config/copr) has a higher priority (preferred, if properly configured)
in copr-cli.

The client side newly uses requests.Session() (and API v3 frontend code
accepts session cookies), so we perform the GSSAPI authentication only
once and not for all the requests.

This is the first part of the effort, we'll continue in issue #1822.

Merges: #1820