frontend: colorize CoprDir-buttons on builds page
This introduces a new KEEP_PR_DIRS_DAYS option, by default set to 40.
When the lastest build in the pull-request CoprDir was submitted more
than 40 days ago, the "filter" button in web-UI will be colored "red"
(as outdated).  When it is older than 1/4*40 (==20) days, the button
will be "orange".

All the buttons have a pop-up title saying how long the directory will
be kept.