frontend: log basic request information for each traceback
I always struggle to figure out how some logged traceback was caused
and how to reproduce it.

This change will probably not be helpful every time but I suspect it
will make the debugging easier in many cases.

Also, I hope that logging username of a person that caused a traceback
won't be considered a privacy breach and user tracking. Having a
contact in case something is notoriously tracebacking for an unknown
reason would be very valuable and could help us improve the service.