frontend, doc: use the same repofile for all Fedora versions including Rawhide
Fix #639
See PR#492

With this patch we will generate the same `baseurl` for all Fedora
versions. Therefore there won't be special `fedora-rawhide-$basearch`
for Rawhide but rather `fedora-$releasever-$basearch`.

When additionally configuring

    CHROOT_NAME_RELEASE_ALIAS = {
	"fedora-35": "fedora-rawhide",
    }

The following URL alias will start to work
http://127.0.0.1:5000/coprs/frostyx/rawhide-repos/repo/fedora-35/
even though technically, there is no `fedora-35-*` chroot in the
database.

Thanks to this, both

    dnf copr enable frostyx/rawhide-repos --releasever=35 --hub=localhost
    dnf copr enable frostyx/rawhide-repos --releasever=rawhide --hub=localhost

works as expected and enable the repository correctly. The only
missing piece is that the following command works

    dnf install foo --refresh --releasever=rawhide

but this one doesn't

    dnf install foo --refresh --releasever=35

AFAIK this is not a regression so it probably shouldn't be a
blocker. But I am for solving it now, if you agree. Do you have any
suggestions what would be ideal solution to link `fedora-35-*` chroots
on the backend to `fedora-rawhide-*`? Maybe some mod_rewrite thingy?