frontend, docker: make template caching configurable
At times, I am confused why my changes in some functions doesn't take
effect, just to realize that they are cached. IMHO we don't want to
cache anything in a local development setup because the caching
mechanism is provided by a third-party package, so we don't need to
work with its code.