frontend: homepage redesign
See #1890

My apologies for a large commit, I started working on the redesign
without any clear idea in what order do I want to change the things,
so they all ended up tangled together.

This commit adds a couple of things to the homepage

1) A header saying something like
   "Copr hosts 16,724 projects from 2,739 Fedora users"
   The string is defined within a `copr-frontend-fedora` specific
   file, and therefore can be easily changed for other instances.

2) A large search bar with instructions how to search by attributes
   such as username, groupname, etc. I used the code from our search
   bar in the top right, moved it to `_helpers.html` and created a
   macro from it, so there is no duplication.

3) A row of cards with some interesting documentation sections.
   They are defined within a `copr-frontend-fedora` specific
   file, and therefore can be easily changed for other instances.

4) If you are logged in, your own pinned projects. For users, that
   didn't customize their pinned projects, we show their most recent
   projects as a fallback.

For non-logged visitors, the pinned projects section is obviously not
displayed. Instead, the vertical space is utilized by our
`welcome.html` that displays simple information how to use Copr. I
plan to rework this box a little in a follow-up PR.

We still show the most recent projects at the bottom of the page. This
will be changed into some more usefull list of projects (e.g. most
popular, editor pick, or something else) in a follow-up PR.