frontend: web-ui: server-side pagination for too-many-packages
When the list of packages is larger than N (1000 currently), instead of
loading all-in-one page and sort by JavaScript - provide a server-side
pagination feature.

This approach is slightly different than what we use in the Builds
server-side pagination -- here we don't work with SQL queries, but with
a list of packages (to optimize the expensive queries).  Therefore I had
to split the get_packages_with_latest_builds_for_dir() method into two
parts, and do the pagination in the first part (get_all_ordered() called
directly from the route method).  We call
get_packages_with_latest_builds_for_dir() also for the API, but the
change was done so API call is not affected.

Follow-up-for: previous-commit
Fixes: #54