frontend: web-ui: paginate monitor page for more than 1000 packages
When the list of packages is larger than N (1000 currently), instead of
loading all-in-one page provide a server-side pagination feature.

This required a rewrite of the plain-text SQL query to SQLAlchemy jargon
to allow the Flask pagination feature.  Having full BuildChroot objects
in hand, in array - it allowed me to rewrite the templates so they are
much more readable.

Fixes: #54